[GraphsUndirectedWeightedSparseGraphTest] - fix unit-texts
